Aggiunti test

+++ src/core/model.ts
@@ -23,18 +23,18 @@ export class Model {
     private readonly TAXES: Taxes;
     private readonly CATEGORIES: Categories;
 
-    private constructor (db: Persistence, users: Users, taxes: Taxes, categories: Categories) {
+    private constructor(db: Persistence, users: Users, taxes: Taxes, categories: Categories) {
         this.DATABASE = db;
-        this.CATEGORIES=categories;
-        this.TAXES=taxes;
-        this.USERS=users;
+        this.CATEGORIES = categories;
+        this.TAXES = taxes;
+        this.USERS = users;
     }
 
-    public static createModel (): Model {
+    public static createModel(): Model {
         return new Model(new Dynamo(), new UsersService(), new TaxesService(), new CategoriesService());
     }
 
-    public static createModelMock (): Model {
+    public static createModelMock(): Model {
         return new Model(new DbMock(), new UsersMock(), new TaxesMock(), new CategoriesMock());
     }
 
@@ -43,8 +43,8 @@ export class Model {
          * @param id id of the product
          * @returns the product with the given id
          */
-    public async getProduct (id: string): Promise<Product> {
-        const PRODUCT_DB: any = await this.DATABASE.get(id);     
+    public async getProduct(id: string): Promise<Product> {
+        const PRODUCT_DB: any = await this.DATABASE.get(id);
         if (PRODUCT_DB == null) {
             console.log("Product " + id + " not found");
             throw new Error("A product with this id doesn't exist");
@@ -55,16 +55,25 @@ export class Model {
         return PRODUCT;
     }
 
-    public async changeStock (id: string, quantity: number, token: string): Promise<boolean> {
+    public async changeStock(id: string, quantity: number, token: string): Promise<boolean> {
         //check if the user is vendor
         const IS_AUTH = await this.USERS.checkUser(token);
-        if (!IS_AUTH){
+        if (!IS_AUTH) {
             throw new Error("invalid token");
         }
-        const PRODUCT= await this.DATABASE.get(id);
+
+        if (!(id && !isNaN(quantity) && (quantity || quantity === 0))) {
+            throw new Error("Missing id or quantity");
+        }
+
+        if (quantity === 0) {
+            return true; // quantità invariata
+        }
+
+        const PRODUCT = await this.DATABASE.get(id);
         PRODUCT.stock = PRODUCT.stock + quantity;
-        const TO_UPDATE = {"stock": PRODUCT.stock}
-        return await  this.DATABASE.update(id, TO_UPDATE);
+        const TO_UPDATE = { "stock": PRODUCT.stock }
+        return await this.DATABASE.update(id, TO_UPDATE);
     }
 
     /**
@@ -72,44 +81,40 @@ export class Model {
          * @param data product data
          * @returns the result of the insertion
          */
-    public async addProduct (data: { [key: string]: any }, token: string):
+    public async addProduct(data: { [key: string]: any }, token: string):
         Promise<boolean> {
         //check if the user is vendor
         const IS_VENDOR = await this.USERS.checkVendor(token);
-        if (!IS_VENDOR){
+        if (!IS_VENDOR) {
             throw new Error("invalid token");
         }
 
-        // handle product image (only if there is an image)
-        if (data.primaryPhoto) {
-            try {
-                const BUCKETNAME = process.env.PRODUCT_IMG_BUCKET;
-                //DATA.image become an URL
-                data.primaryPhoto = await s3.uploadImage(data.primaryPhoto, BUCKETNAME);
-                if (data.secondaryPhotos) {
-                    for (let i = 0; i < data.secondaryPhotos.length; i++) {
-                        data.secondaryPhotos[i] = 
-                        await s3.uploadImage(data.secondaryhotos[i], BUCKETNAME);
-                    }
-                }
-            }
-            catch (err) {
-                console.log(err.message)
-                return false;
-            }
+        if (!data) {
+            throw new Error("New product info required");
         }
 
         //validate and add to db
         const VALID = this.AJV.validate("schemas/products.json#/insertProduct", data);
-        if (VALID) {
-            data.id = uuidv4();
-            const RES: Promise<boolean> = this.DATABASE.write(JSON.parse(JSON.stringify(data)));
-            return RES;
+        if (!VALID) {
+            throw new Error("Product does not match the schema of required attributes");
         }
-        else {
-            // i dati in ingresso per creare l'oggetto non rispettano lo json-schema
-            return false;
+
+        // handle product image (only if there is an image)
+        if (data.primaryPhoto) {
+            const BUCKETNAME = process.env.PRODUCT_IMG_BUCKET;
+            //DATA.image become an URL
+            data.primaryPhoto = await s3.uploadImage(data.primaryPhoto, BUCKETNAME);
+            if (data.secondaryPhotos) {
+                for (let i = 0; i < data.secondaryPhotos.length; i++) {
+                    data.secondaryPhotos[i] =
+                        await s3.uploadImage(data.secondaryPhotos[i], BUCKETNAME);
+                }
+            }
         }
+
+        data.id = uuidv4();
+        const RES: Promise<boolean> = this.DATABASE.write(JSON.parse(JSON.stringify(data)));
+        return RES;
     }
 
     /**
@@ -118,47 +123,56 @@ export class Model {
          * @param DATA data to be updated
          * @returns the result of the query
         */
-    public async updateProduct (PRODUCT_ID: string, DATA: {[key:string]: any}, token: string):
+    public async updateProduct(PRODUCT_ID: string, DATA: { [key: string]: any }, token: string):
         Promise<boolean> {
         //check if the user is vendor
         const IS_VENDOR = await this.USERS.checkVendor(token);
-        if (!IS_VENDOR){
+        if (!IS_VENDOR) {
             throw new Error("invalid token");
         }
+
+        if(!PRODUCT_ID){
+            throw new Error("Product id not valid");
+        }
+
+        const VALID = this.AJV.validate("schemas/products.json#/editProduct", DATA);
+        if (!VALID) {
+            throw new Error("Product does not match the schema of required attributes");
+        }
+
         //image replacing, if needed
         if (DATA['primaryPhoto']) {
             const BUCKETNAME = process.env.PRODUCT_IMG_BUCKET;
             s3.deleteFile(BUCKETNAME, DATA['primaryPhoto'].key)
             DATA['primaryPhoto'] = await s3.uploadImage(DATA['primaryPhoto'], BUCKETNAME);
         }
+
         if (DATA['secondaryPhotos']) {
             const BUCKETNAME = process.env.PRODUCT_IMG_BUCKET;
             s3.deleteFile(BUCKETNAME, DATA[''].key)
             DATA['primaryPhoto'] = await s3.uploadImage(DATA['primaryPhoto'], BUCKETNAME);
         }
 
-        const VALID = this.AJV.validate("schemas/products.json#/editProduct", DATA);
-        if (VALID) {
-            const PRODUCT = await this.DATABASE.update(PRODUCT_ID, DATA);
-            return PRODUCT;
-        }
-        else {
-            // i dati in ingresso per modificare l'oggetto non rispettano lo json-schema
-            return false;
-        }
+        const PRODUCT = await this.DATABASE.update(PRODUCT_ID, DATA);
+        return PRODUCT;
     }
 
-    public async deleteProduct (PRODUCT_ID: string, token: string):
-    Promise<boolean> {
-    //check if the user is vendor
+    public async deleteProduct(PRODUCT_ID: string, token: string):
+        Promise<boolean> {
+        //check if the user is vendor
         const IS_VENDOR = await this.USERS.checkVendor(token);
-        if (!IS_VENDOR){
+        if (!IS_VENDOR) {
             throw new Error("invalid token");
         }
+
+        if(!PRODUCT_ID){
+            throw new Error("Product id not valid");
+        }
+
         return await this.DATABASE.delete(PRODUCT_ID);
     }
 
-    public async getProducts (category: string, minPrice: number, maxPrice: number, sorting: string): Promise<any>{
+    public async getProducts(category: string, minPrice: number, maxPrice: number, sorting: string): Promise<any> {
         let sortingAsc = null;
         if (sorting) {
             if (sorting == "asc")
@@ -172,21 +186,7 @@ export class Model {
             throw new Error("error while retrieving products from db")
         const PRODUCTS = [];
         //microservices calls
-        for (let i = 0; i < PRODUCTS_DB.length; i++){
-            const PRODUCT= await this.createProduct(PRODUCTS_DB[i]);
-            console.log("Product " + i + ": " + JSON.stringify(PRODUCT));
-            PRODUCTS.push(PRODUCT);
-        }
-        return PRODUCTS;
-    }
-
-    public async getHomeProducts (): Promise<any>{
-        const PRODUCTS_DB = await this.DATABASE.getProductsHome();
-        if(!PRODUCTS_DB)
-            throw new Error("error while retrieving products from db")
-        const PRODUCTS = [];
-        //microservices calls
-        for(let i = 0; i < PRODUCTS_DB.length; i++){
+        for (let i = 0; i < PRODUCTS_DB.length; i++) {
             const PRODUCT = await this.createProduct(PRODUCTS_DB[i]);
             console.log("Product " + i + ": " + JSON.stringify(PRODUCT));
             PRODUCTS.push(PRODUCT);
@@ -194,10 +194,26 @@ export class Model {
         return PRODUCTS;
     }
 
-    public async getProductsByName (name:string): Promise<any>{
-        
-        const PRODUCTS_DB = await this.DATABASE.getProductsByName(name); // From array of array to single array
-        
+    public async getHomeProducts(): Promise<any> {
+        const PRODUCTS_DB = await this.DATABASE.getProductsHome();
+        if (!PRODUCTS_DB)
+            throw new Error("error while retrieving products from db")
+        const PRODUCTS = [];
+        //microservices calls
+        for (let i = 0; i < PRODUCTS_DB.length; i++) {
+            const PRODUCT = await this.createProduct(PRODUCTS_DB[i]);
+            console.log("Product " + i + ": " + JSON.stringify(PRODUCT));
+            PRODUCTS.push(PRODUCT);
+        }
+        return PRODUCTS;
+    }
+
+    public async getProductsByName(name: string): Promise<any> {
+        if (!name) {
+            throw new Error("search string can not be empty");
+        }
+        const PRODUCTS_DB = await this.DATABASE.getProductsByName(name);
+
         if (!PRODUCTS_DB)
             throw new Error("error while retrieving products from db")
         const PRODUCTS = [];
@@ -211,13 +227,13 @@ export class Model {
     }
 
     //PRIVATE METHODS
-    private async createProduct (product: any): Promise<Product> {
+    private async createProduct(product: any): Promise<Product> {
         //tax
         const TAX = await this.TAXES.getTax(product.tax);
-        const TAX_VALUE = TAX["value"];       
+        const TAX_VALUE = TAX?.value;
         product.tax = TAX_VALUE;
         //category
-        const CATEGORY= await this.CATEGORIES.getCategoryName(product.category);
+        const CATEGORY = await this.CATEGORIES.getCategoryName(product.category);
         product.category = CATEGORY;
         return new Product(product);
     }
